requirements.txt
aiobotocore==1.4.2
aiofiles==0.8.0
aiohttp==3.7.4.post0
aioitertools==0.8.0
aiomysql==0.0.21
APScheduler==3.8.1
asgiref==3.4.1
astroid==2.11.5
async-lru==1.0.2
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
bidict==0.21.4
boto3==1.17.3
botocore==1.20.106
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.1.0
click==7.1.2
coverage==6.3.2
cssselect==1.1.0
cssutils==1.0.2
db-dtypes==1.0.2
dill==0.3.5.1
dnspython==2.0.0
elasticsearch==6.8.2
email-validator==1.1.2
et-xmlfile==1.1.0
fastapi==0.62.0
fastapi-socketio==0.0.9
fastapi-utils==0.2.1
filelock==3.4.0
google-api-core==2.8.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
googletrans==4.0.0rc1
greenlet==1.1.1
grpcio==1.47.0
grpcio-status==1.47.0
gspread==4.0.1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.2
httptools==0.1.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==1.1.1
isort==5.10.1
jmespath==0.10.0
lazy-object-proxy==1.7.1
lxml==4.6.2
mccabe==0.7.0
multidict==5.2.0
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.5.5.62
openpyxl==3.0.9
orjson==3.6.7
packaging==20.9
pandas==1.3.4
Pillow==8.4.0
platformdirs==2.5.2
pluggy==0.13.1
premailer==3.7.0
proto-plus==1.20.6
protobuf==3.19.1
py==1.10.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pyfcm==1.5.4
PyJWT==2.0.0
pylint==2.13.9
PyMySQL==0.9.3
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
pytest-cov==3.0.0
python-dateutil==2.8.1
python-dotenv==0.19.0
python-engineio==4.3.1
python-gettext==4.0
python-multipart==0.0.5
python-socketio==5.5.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.3.4
shortio==0.1.0
sio==0.1.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.22
starlette==0.13.6
starlette-context==0.3.3
toml==0.10.2
tomli==2.0.1
tqdm==4.62.3
tweepy==4.10.0
typing-extensions==3.10.0.0
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
uritemplate==4.1.1
urllib3==1.26.2
uvicorn==0.15.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
wrapt==1.13.2
XlsxWriter==3.0.3
yagmail==0.14.245
yarl==1.7.0
aioredis-py==2.0.1